requirements.txt
amqp==2.6.1
billiard==3.6.3.0
bootstrap-admin==0.4.4
celery==4.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-proxy==1.2.1
django-redis==4.12.1
django-simpleui==3.8
django-suit==0.2.28
django-timezone-field==4.0
djangorestframework==3.11.1
dnspython==1.16.0
docutils==0.16
et-xmlfile==1.0.1
eventlet==0.26.1
flower==0.9.7
greenlet==0.4.16
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.17.0
kombu==4.6.11
lockfile==0.12.2
MarkupSafe==1.1.1
monotonic==1.5
mysql==0.0.3
mysqlclient==2.0.1
nacos-sdk-python==0.1.5
numpy==1.19.5
openapi-codec==1.3.2
openpyxl==3.0.4
pika==1.3.1
prometheus-client==0.8.0
psutil==5.8.0
PyMySQL==1.0.2
python-crontab==2.5.1
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
pandas==1.1.5
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tornado==6.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
